requirements.txt
anthropic>=0.40.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.35.0
